+++ src/main/java/me/lkp111138/deebot/commands/ConfigCommand.java
@@ -49,7 +49,7 @@ public class ConfigCommand implements Command {
             case group:
             case supergroup:
                 long gid = msg.chat().id();
-                int uid = msg.from().id();
+                long uid = msg.from().id();
                 try (PreparedStatement stmt = Main.getConnection().prepareStatement("SELECT fry, collect_place FROM `groups` where gid=?")) {
                     stmt.setLong(1, gid);
                     ResultSet rs = stmt.executeQuery();
