+++ src/main/java/me/lkp111138/deebot/commands/HelpCommand.java
@@ -9,7 +9,7 @@ import me.lkp111138.deebot.translation.Translation;
 public class HelpCommand implements Command {
     @Override
     public void respond(TelegramBot bot, Message msg, String[] args) {
-        int sender = msg.from().id();
+        long sender = msg.from().id();
         SendMessage req = new SendMessage(sender, Translation.get(null).HELP()).parseMode(ParseMode.Markdown);
         bot.execute(req);
     }
